Share cloud TTS audio lifecycle
